# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamGuard API - Batch IPTV stream health checker and playlist manager
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ApiMessage {
  message: string;
}

export type PlaylistSourceType = typeof PlaylistSourceType[keyof typeof PlaylistSourceType];


export const PlaylistSourceType = {
  text: 'text',
  url: 'url',
  file: 'file',
} as const;

export interface Playlist {
  id: number;
  name: string;
  sourceType: PlaylistSourceType;
  /** @nullable */
  sourceUrl?: string | null;
  entryCount: number;
  duplicatesFound: number;
  parseWarnings: string[];
  groups: string[];
  createdAt: string;
}

export type PlaylistInputSourceType = typeof PlaylistInputSourceType[keyof typeof PlaylistInputSourceType];


export const PlaylistInputSourceType = {
  text: 'text',
  url: 'url',
  file: 'file',
} as const;

export interface PlaylistInput {
  name: string;
  sourceType: PlaylistInputSourceType;
  /**
     * Raw M3U/M3U8 text (for sourceType=text) or base64 file (for sourceType=file)
     * @nullable
     */
  content?: string | null;
  /**
     * Remote URL to fetch (for sourceType=url)
     * @nullable
     */
  url?: string | null;
}

export interface Channel {
  id: number;
  playlistId: number;
  /** @nullable */
  tvgId?: string | null;
  /** @nullable */
  tvgName?: string | null;
  /** @nullable */
  tvgLogo?: string | null;
  /** @nullable */
  groupTitle?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  country?: string | null;
  url: string;
  position: number;
}

export interface ChannelPage {
  channels: Channel[];
  total: number;
  page: number;
  limit: number;
}

export interface JobSettings {
  concurrency?: number;
  timeoutMs?: number;
  retryCount?: number;
  autoProbe?: boolean;
  userAgents?: string[];
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  queued: 'queued',
  running: 'running',
  paused: 'paused',
  completed: 'completed',
  cancelled: 'cancelled',
  failed: 'failed',
} as const;

export interface Job {
  id: number;
  playlistId: number;
  playlistName: string;
  status: JobStatus;
  settings: JobSettings;
  total: number;
  checked: number;
  live: number;
  dead: number;
  geoblocked: number;
  suspicious: number;
  pending: number;
  /** @nullable */
  etaSeconds?: number | null;
  /** @nullable */
  avgCheckMs?: number | null;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface JobInput {
  playlistId: number;
  settings?: JobSettings;
}

export interface CategoryCount {
  category: string;
  total: number;
  live: number;
  dead: number;
  geoblocked: number;
  suspicious: number;
  pending: number;
}

export interface JobSummary {
  jobId: number;
  status: string;
  total: number;
  checked: number;
  live: number;
  dead: number;
  geoblocked: number;
  suspicious: number;
  pending: number;
  livePercent: number;
  deadPercent: number;
  /** @nullable */
  avgCheckMs?: number | null;
  /** @nullable */
  etaSeconds?: number | null;
  topCategories: CategoryCount[];
  progressPercent: number;
}

export type ChannelResultStatus = typeof ChannelResultStatus[keyof typeof ChannelResultStatus];


export const ChannelResultStatus = {
  live: 'live',
  dead: 'dead',
  geoblocked: 'geoblocked',
  suspicious: 'suspicious',
  pending: 'pending',
} as const;

export interface ProbeData {
  /** @nullable */
  videoCodec?: string | null;
  /** @nullable */
  audioCodec?: string | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  framerate?: number | null;
  /** @nullable */
  bitrate?: number | null;
  /** @nullable */
  container?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  mislabeled?: boolean;
  /** @nullable */
  mislabelReason?: string | null;
}

export interface ChannelResult {
  id: number;
  jobId: number;
  channelId: number;
  /** @nullable */
  tvgName?: string | null;
  /** @nullable */
  tvgLogo?: string | null;
  url: string;
  /** @nullable */
  category?: string | null;
  status: ChannelResultStatus;
  /** @nullable */
  httpStatus?: number | null;
  /** @nullable */
  responseTimeMs?: number | null;
  /** @nullable */
  redirectCount?: number | null;
  /** @nullable */
  tlsValid?: boolean | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  manifestValid?: boolean | null;
  /** @nullable */
  failureReason?: string | null;
  probeData?: ProbeData;
  /** @nullable */
  checkedAt?: string | null;
}

export interface JobDiff {
  jobId: number;
  compareJobId: number;
  newlyDead: ChannelResult[];
  newlyLive: ChannelResult[];
  unchanged: number;
  summary: string;
}

export interface ResultPage {
  results: ChannelResult[];
  total: number;
  page: number;
  limit: number;
}

export interface ResultUpdate {
  /** @nullable */
  category?: string | null;
}

export interface ProbeRequest {
  resultIds: number[];
}

export interface Settings {
  defaultConcurrency: number;
  defaultTimeoutMs: number;
  defaultRetryCount: number;
  maxConcurrency: number;
  perHostConcurrency: number;
  autoProbeDefault: boolean;
  ffprobePath: string;
}

export interface SettingsInput {
  /** @nullable */
  defaultConcurrency?: number | null;
  /** @nullable */
  defaultTimeoutMs?: number | null;
  /** @nullable */
  defaultRetryCount?: number | null;
  /** @nullable */
  maxConcurrency?: number | null;
  /** @nullable */
  perHostConcurrency?: number | null;
  /** @nullable */
  autoProbeDefault?: boolean | null;
  /** @nullable */
  ffprobePath?: string | null;
}

export type GetPlaylistChannelsParams = {
page?: number;
limit?: number;
group?: string;
search?: string;
};

export type GetJobResultsParams = {
page?: number;
limit?: number;
status?: string;
category?: string;
search?: string;
sortBy?: string;
sortDir?: string;
};

export type ExportJobParams = {
format: string;
/**
 * Comma-separated status filter e.g. "live,suspicious"
 */
status?: string;
category?: string;
};