# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamGuard API - Batch IPTV stream health checker and playlist manager
 * OpenAPI spec version: 0.1.0
 */

export interface Channel {
  id: number;
  playlistId: number;
  /** @nullable */
  tvgId?: string | null;
  /** @nullable */
  tvgName?: string | null;
  /** @nullable */
  tvgLogo?: string | null;
  /** @nullable */
  groupTitle?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  country?: string | null;
  url: string;
  position: number;
}