# lib/db/src/schema/playlists.ts
import { pgTable, serial, text, integer, timestamp, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const playlistsTable = pgTable("playlists", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  sourceType: text("source_type").notNull(), // text | url | file
  sourceUrl: text("source_url"),
  entryCount: integer("entry_count").notNull().default(0),
  duplicatesFound: integer("duplicates_found").notNull().default(0),
  parseWarnings: jsonb("parse_warnings").notNull().$type<string[]>().default([]),
  groups: jsonb("groups").notNull().$type<string[]>().default([]),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertPlaylistSchema = createInsertSchema(playlistsTable).omit({ id: true, createdAt: true });
export type InsertPlaylist = z.infer<typeof insertPlaylistSchema>;
export type Playlist = typeof playlistsTable.$inferSelect;